harn-lint: split tests.rs into rule-family submodules
Closes #44. tests.rs was 2290 lines of 122 #[test] fns spanning every
rule family. Splitting makes it tractable to locate and extend a
rule's coverage without scrolling through unrelated tests.

Shared helpers (lint_source, has_rule, count_rule, get_fix,
apply_fixes, lint_with_require_header) live only in tests/mod.rs,
exposed to submodules via pub(super).

All 122 tests still run with the same names and pass/fail outcomes;
largest submodule file is 313 lines, well under the 800-line cap.
